+++ .dir-locals.el
@@ -1,0 +1,4 @@
+;;; Directory Local Variables            -*- no-byte-compile: t -*-
+;;; For more information see (info "(emacs) Directory Variables")
+((nil . ((projectile-project-test-cmd . "hatch run test"))))
+